Fixed all: warning C4244: 'initializing' : conversion from '...' to '...', possible loss of data errors
- Probleem in MainEntityContainer en NightLayer was de FLOOR() methode
welke een double returned terwijl we maar een INT nodig hadden (geen
comma nodig)

+++ Bermuda/Bermuda/Game.cpp
@@ -50,7 +50,7 @@ void Game::gameLoop(GameStateManager* gsm) {
 		//Take current Time
 		QueryPerformanceCounter(&currentTime);
 		//Calculate difference (previousTime - currentTime)
-		long updateLength = (currentTime.QuadPart - previousTime.QuadPart) * 1000.0 / frequency.QuadPart;
+		long updateLength = static_cast<long>((currentTime.QuadPart - previousTime.QuadPart) * 1000.0 / frequency.QuadPart);
 		
 		//Previous time = current time
 		previousTime = currentTime;
